pyproject.toml
[project]
name = "attendance-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepface>=0.0.98",
    "fastapi>=0.128.5",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "passlib[bcrypt]>=1.7.4",
    "pytelegrambotapi>=4.30.0",
    "python-dotenv==1.0.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.46",
    "tf-keras>=2.20.1",
    "uvicorn>=0.40.0",
]
